Add more macOS guards

+++ chef/cookbooks/cpe_nightly_reboot/recipes/default.rb
@@ -10,4 +10,6 @@
 # of patent rights can be found in the PATENTS file in the same directory.
 #
 
+return unless node.macos?
+
 cpe_nightly_reboot 'Create nightly reboot'
